Controllo forno funzionante, Controllo velocita conveyor funzionante

+++ CodeProject/src/main/java/org/example/MainProject/CookiesFactory.java
@@ -30,7 +30,7 @@ public class CookiesFactory {
         return conveyor.isSignalReady();
     }
 
-    public int getConveyorRPM() {
+    public double getConveyorRPM() {
         return conveyor.getRpm();
     }
